vectorization/requirements.txt
requests==2.31.0
Flask==3.0.2
python-dotenv==1.0.1
psycopg2==2.9.9
llama-index==0.10.12
llama-index-vector-stores-postgres==0.1.2
docx2txt==0.8
pypdf==4.0.1
pgvector==0.2.5
psycopg2-binary==2.9.9
torch==2.2.1
transformers==4.38.1
python-pptx==0.6.23
pillow==10.2.0
openpyxl==3.1.2